+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 2.0.1 (TBD)
 
-* TBD
+* geopackage-android version updated to 2.0.1
+* Visible flag support for Option typed (pre map) Google Map Shapes
+* Z Index support for Google Map Shapes
 
 ## [2.0.0](https://github.com/ngageoint/geopackage-android-map/releases/tag/2.0.0) (11-20-2017)